core module and test working on Linux (Ubuntu)

+++ Engine/Source/Core/Public/Misc/CoreLiterals.h
@@ -1,22 +1,22 @@
 #pragma once
 
 
-constexpr SIZE_T operator"" _KB(SIZE_T Value)
+constexpr SIZE_T operator"" _KB(unsigned long long Value)
 {
     return Value * 1024;
 }
 
-constexpr SIZE_T operator"" _MB(SIZE_T Value)
+constexpr SIZE_T operator"" _MB(unsigned long long Value)
 {
     return Value * (1024 * 1024);
 }
 
-constexpr SIZE_T operator"" _GB(u64 Value)
+constexpr SIZE_T operator"" _GB(unsigned long long Value)
 {
     return Value * (1024 * 1024 * 1024);
 }
 
-constexpr SIZE_T operator"" _Bytes(SIZE_T Value)
+constexpr SIZE_T operator"" _Bytes(unsigned long long Value)
 {
     return Value;
 }
